+++ cquill/examples/session_connection.rs
@@ -1,12 +1,20 @@
+use std::sync::Arc;
+
 use cquill::*;
+use scylla::client::session_builder::SessionBuilder;
 
 #[tokio::main]
 async fn main() {
     let cql_dir = std::path::PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").unwrap())
         .join("examples")
         .join("cql");
+    let session = SessionBuilder::new()
+        .known_node("127.0.0.1")
+        .build()
+        .await
+        .unwrap();
     let opts = MigrateOpts {
-        connection_opts: None,
+        connection_opts: Some(ConnectionOpts::Session(Arc::new(session))),
         cql_dir,
         history_keyspace: None,
         history_table: None,